pyproject.toml
[project]
name = "okta-agentcore-auth-service"
version = "0.1.0"
description = "Minimal authentication service for Okta client-credentials and AWS AgentCore Gateway forwarding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<0.29.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "pyjwt[crypto]>=2.10.0,<3.0.0",
    "uvicorn>=0.34.0,<0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.25.0,<0.26.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
